tendermint-proxy: 🎈 hoist `TendermintProxy` to `lib.rs`
there are only two files in this library. rather than reëxporting the
type in the `lib.rs`, just define it there and import it where we use
it.

this puts the important, core type provided by the library right at the
top of the reader's entrypoint.
